+++ laboratorium.cabal
@@ -26,6 +26,9 @@ library
     Query.Retrograde
     Query.Crossing
     Query.Transit
+    Query.Transit.Types
+    Query.Transit.Interplanetary
+    Query.Transit.Lunar
     Query.Common
     Query.Aggregate
     Query.Streaming
